# docs/swagger.yaml
basePath: /api
definitions:
  edutalks_internal_models.Article:
    properties:
      authorId:
        type: integer
      bodyHtml:
        type: string
      createdAt:
        type: string
      id:
        type: integer
      isPublished:
        type: boolean
      publishedAt:
        type: string
      summary:
        type: string
      tags:
        items:
          type: string
        type: array
      title:
        type: string
      updatedAt:
        type: string
    type: object
  edutalks_internal_models.CreateArticleRequest:
    properties:
      bodyHtml:
        example: <p>Контент</p>
        type: string
      isPublished:
        type: boolean
      publish:
        type: boolean
      summary:
        example: Короткое описание для превью
        type: string
      tags:
        example:
        - go
        - backend
        - markdown
        items:
          type: string
        type: array
      title:
        example: Как писать middleware в Go
        type: string
    type: object
...
  internal_handlers.SetPublishBody:
    properties:
      publish:
        type: boolean
    type: object
  internal_handlers.changeReq:
    properties:
      new_password:
        type: string
      old_password:
        type: string
    type: object
...
  internal_handlers.forgotReq:
    properties:
      email:
        type: string
    type: object
...
  internal_handlers.resetReq:
    properties:
      new_password:
        type: string
      token:
        type: string
    type: object
...
host: edutalks.ru
info:
  contact:
    email: support@edutalks.ru
    name: EduTalks Support
    url: https://edutalks.ru
  description: Документация API Edutalks (регистрация, логин, токены, статьи, логи
    и т.д.).
  title: Edutalks API
  version: "1.0"
paths:
  /api/admin/articles:
    post:
      consumes:
      - application/json
      - multipart/form-data
      - application/x-www-form-urlencoded
      parameters:
      - description: Данные статьи
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/edutalks_internal_models.CreateArticleRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/edutalks_internal_models.Article'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Создать статью
      tags:
      - articles
  /api/admin/articles/{id}:
    delete:
      parameters:
      - description: ID статьи
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: no content
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Удалить статью
      tags:
      - articles
    patch:
      consumes:
      - application/json
      - multipart/form-data
      - application/x-www-form-urlencoded
      parameters:
      - description: ID статьи
        in: path
        name: id
        required: true
        type: integer
      - description: Данные статьи
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/edutalks_internal_models.CreateArticleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/edutalks_internal_models.Article'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Обновить статью
      tags:
      - articles
  /api/admin/articles/{id}/publish:
    patch:
      consumes:
      - application/json
      parameters:
      - description: ID статьи
        in: path
        name: id
        required: true
        type: integer
      - description: Флаг публикации
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_handlers.SetPublishBody'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/edutalks_internal_models.Article'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Установить публикацию статьи
      tags:
      - articles
  /api/admin/articles/preview:
    post:
      consumes:
      - application/json
      parameters:
      - description: Сырый HTML статьи
        in: body
        name: body
        required: true
        schema:
          additionalProperties:
            type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Предпросмотр статьи
      tags:
      - articles
...
  /api/articles:
    get:
      parameters:
      - description: Номер страницы
        in: query
        name: page
        type: integer
      - description: Размер страницы
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/edutalks_internal_models.Article'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Список статей
      tags:
      - articles
  /api/articles/{id}:
    get:
      parameters:
      - description: ID статьи
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/edutalks_internal_models.Article'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Получить статью по ID
      tags:
      - articles
...
  /api/password/change:
    post:
      consumes:
      - application/json
      description: Смена пароля по старому паролю. Требуется JWT-токен.
      parameters:
      - description: Старый и новый пароль
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/internal_handlers.changeReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Смена пароля (авторизованный пользователь)
      tags:
      - password
  /api/password/forgot:
    post:
      consumes:
      - application/json
      description: Отправляет письмо со ссылкой для сброса пароля. Ответ всегда одинаковый,
        даже если e-mail не найден.
      parameters:
      - description: Email пользователя
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/internal_handlers.forgotReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Запрос восстановления пароля
      tags:
      - password
  /api/password/reset:
    post:
      consumes:
      - application/json
      description: Устанавливает новый пароль по токену из письма.
      parameters:
      - description: Токен и новый пароль
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/internal_handlers.resetReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Сброс пароля по токену
      tags:
      - password
...
schemes:
- https
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"